finished first padding test (and maybe second padding test?)
padding seems to work. first padding test definitely works. for some reason, the second padding test works if you do that test individually (./gnu.exe 5), but doesn't work if you run all the tests (./gnu.exe). this doesn't seem to be a consequence of the seg fault in the later test, because i tested to return after the 5th test in main, but it still happens.